% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsContinuousAccumulationsMatrix.R
\name{continuousAssessment}
\alias{continuousAssessment}
\title{USGS Continuous Oil and Gas Assessment}
\usage{
continuousAssessment(auMC, auType, auProbability, auAreaProductive,
  auAreaDrainage, auPercAreaUntested, auPercAreaSweet, auPercFutureSS, auEURss,
  auPercFutureNS = NULL, auEURns = NULL, auGOR = NULL, auNGLGR = NULL,
  auLGR = NULL, year)
}
\arguments{
\item{auMC}{number of MC iterations to perform}

\item{auType}{type of assessment, string specifier of 'oil' or 'gas'}

\item{auProbability}{Probability: as a decimal percentage}

\item{auAreaProductive}{Productive area of accumulation: \code{c(min,mode,max)} [acres]}

\item{auAreaDrainage}{Uncertainty about average drainage area of wells: \code{c(min,mode,max)} [acres]}

\item{auPercAreaUntested}{Percentage of total assessment-unit area that is untested: \code{c(min,mode,max)} [percentage]}

\item{auPercAreaSweet}{Percentage of untested assessment-unit area in sweet spots: \code{c(min,mode,max)} [percentage]}

\item{auPercFutureSS}{Future success ratio for sweet spots: \code{c(min,mode,max)} [percentage]}

\item{auEURss}{Uncertainty about sweet spot average EUR (MMBO for oil; BCFG for gas): \code{c(min,med,max)} [MMBO or BCFG]}

\item{auPercFutureNS}{Future success ratio for non-sweet spots: \code{c(min,mode,max)} [percentage]}

\item{auEURns}{Uncertainty about non-sweet spot average EUR (MMBO for oil; BCFG for gas): \code{c(min,med,max)} [MMBO or BCFG]}

\item{auGOR}{Gas/oil ratio (CFG/BO): \code{c(min,mode,max)} [CFG/BO]}

\item{auNGLGR}{NGL/gas ratio (BNGL/MMCFG): \code{c(min,mode,max)} [BNGL/MMCFG]}

\item{auLGR}{Liquids/gas ratio (BLIQ/MMCFG): \code{c(min,mode,max)} [BLIQ/MMCFG]}

\item{year}{Year [XXXX] of factsheet publication of assessment numbers.}
}
\value{
A list where each row represents an MC iteration variable combination while the columns contain
 the sample distributions. The column names are as follows
\itemize{
\item\code{DrainageArea} \item\code{EURsweet} \item\code{EURnonSweet} \item\code{ProductiveArea} \item\code{PercAreaUntested}
\item\code{PercAreaSweet} \item\code{PercFutureSweet} \item\code{PercFutureNonSweet} \item\code{untestedArea} \item\code{SweetArea}
\item\code{NonSweetArea} \item\code{NumSweetWells} \item\code{NumNonSweetWells} \item\code{SweetAccumulation} \item\code{NonSweetAccumulation}
\item\code{totalAccumulation}
}
Depending on the type of assessment, also returns (for 'Gas')
\itemize{
\item\code{LGR} \item\code{SweetNGLinGas} \item\code{NonSweetNGLinGas} \item\code{totalNGLinGas}
}
or returns (for 'Oil')
\itemize{
\item\code{GOR} \item\code{NGLGR} \item\code{SweetGasinOil} \item\code{NonSweetGasinOil}
\item\code{totalGasInOil} \item\code{SweetNGLinOil} \item\code{NonSweetNGLinOil} \item\code{totalNGLinOil}
}
}
\description{
Performs USGS assessment summarized in Charpentier (2010) as a vectorized Monte Carlo.
}
\details{
If the AU is to be assessed as entirely sweet spot, simply omit the non-sweet spot variables.
Due to the nature of random sampling, the simulation is vectorized and limited only by the available storage.
}
\note{
Edited by CDMartinez 24 Nov 15.
As of July 2016, the standard 'z-score' for the EUR distribution is implemented as 2.326 (99\%) in
 \code{\link{enforceRankCorrelation}}. Assigning \code{year} <= 1 will assume zscore of 3.09 (99.9\%)
 as used in \code{\link{conventionalAssessment}} EUR distributions since approximately 2013.
}
\examples{
OGasmt <- continuousAssessment(auMC = 50000,
auType = 'Oil',
auProbability = 1,
auAreaProductive = c(2800000,3100000,3400000),
auAreaDrainage = c(320,400,600),
auPercAreaUntested = c(80,87,91),
auPercAreaSweet = c(24,29,70),
auPercFutureSS = c(98,99,100),
auEURss = c(0.225,0.25,0.325),
auPercFutureNS = c(80,90,95),
auEURns = c(0.075,0.15,0.25),
auGOR = c(500,1000,1500),
auNGLGR = c(35,85,115),
year = 2013)
round(continuousAssessmentSummary(OGasmt$risked))
}
\references{
Charpentier, Ronald R, and Troy Cook, 2010, Improved USGS Methodology for Assessing Continuous Petroleum
Resources. U.S. Geological Survey Data Series 547, 2: 22.
}
\author{
Created by CDMartinez 10 Nov 15
}
