% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsContinuousAccumulationsMatrix.R
\name{continuousAssessmentSummary}
\alias{continuousAssessmentSummary}
\title{Print USGS Continuous Assessment Summary}
\usage{
continuousAssessmentSummary(dataList)
}
\arguments{
\item{dataList}{The list returned from \code{\link{continuousAssessment}}}
}
\value{
Printed table in the console
}
\description{
Returns a matrix table of the accumulation mean and fractiles typically published in USGS Factsheets.
}
\note{
Edited by CDMartinez  1 Nov 15
}
\examples{
OGasmt <- continuousAssessment(auMC = 50000,
auType = 'Oil',
auProbability = 1,
auAreaProductive = c(2800000,3100000,3400000),
auAreaDrainage = c(320,400,600),
auPercAreaUntested = c(80,87,91),
auPercAreaSweet = c(24,29,70),
auPercFutureSS = c(98,99,100),
auEURss = c(0.225,0.25,0.325),
auPercFutureNS = c(80,90,95),
auEURns = c(0.075,0.15,0.25),
auGOR = c(500,1000,1500),
auNGLGR = c(35,85,115),
year = 2013)
round(continuousAssessmentSummary(OGasmt$risked))
}
\author{
Created by CDMartinez  1 Nov 15
}
