% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsConventionalAccumulations.R
\name{conventionalAssessment}
\alias{conventionalAssessment}
\title{USGS Conventional Oil and Gas Assessment}
\usage{
conventionalAssessment(auMC, auType, auCharge, auRocks, auTiming, auFieldNum,
  auFieldSize, auGOR = NULL, auNGLGR = NULL, auLGR = NULL, auOGR = NULL,
  year)
}
\arguments{
\item{auMC}{number of simulations}

\item{auType}{'Oil' or 'Gas'}

\item{auCharge}{charge probability [between 0 and 1]}

\item{auRocks}{rock probability [between 0 and 1]}

\item{auTiming}{timing probability [between 0 and 1]}

\item{auFieldNum}{Number of possible undiscovered accumulations; \code{c(min,mode,max)} [unitless]}

\item{auFieldSize}{Size of undiscovered accumulations; \code{c(min,mode,max)} [unitless]}

\item{auGOR}{Gas to oil ratio; \code{c(min,mode,max)} [CFG/BO]}

\item{auNGLGR}{Natural gas liquids to gas ratio; \code{c(min,mode,max)} [BNGL/MMCFG]}

\item{auLGR}{Liquids to gas ratio; \code{c(min,mode,max)} [BLIQ/MMCFG]}

\item{auOGR}{Oil to gas ratio; \code{c(min,mode,max)} [BO/MMCFG]}

\item{year}{The year assessment was published. Mathematical implementation varies based on year.}
}
\value{
A list containing three items: list of \code{risked} outputs, list of \code{unrisked}
outputs, and vector of Field Sizes (\code{mcFieldSize}). Each row in the
\code{risked} and \code{unrisked} lists represent a singel Monte Carlo iteration variable
combination while the columns contain the sample distributions.

The \code{risked} and \code{unrisked} column names for 'Oil' assessment are as follows:
\itemize{
\item\code{mcProbability} \item\code{totalAccumulation} \item\code{totalGOR}
\item\code{totalNGLGR} \item\code{mcNumOil} \item\code{mcGOR} \item\code{mcNGLGR} \item\code{mcMEL}
}
The \code{risked} and \code{unrisked} column names for 'Gas' assessment are as follows:
\itemize{
\item\code{mcProbability} \item\code{totalAccumulation} \item\code{totalOGR}
\item\code{totalLGR} \item\code{mcNumGas} \item\code{mcOGR} \item\code{mcLGR} \item\code{mcMEL}
}
}
\description{
Performs USGS assessment summarized in Schmoker (2005) as a vectorized Monte Carlo.
}
\details{
Due to the nature of random sampling, the simulation is vectorized and limited only by the available storage. The assessment year determines slightly different implementations:
\enumerate{
\item \code{year < 2014} uses shifted-truncated lognormal distributions for size of fields only. Triangular for number of fields and coproducts.
\item \code{year >= 2014} uses shifted-truncated lognormal distributions for size of fields and number of fields. Triangular for coproducts.
\item \code{year < 2013} uses distribution z-score of 2.326 (99\% of the distribution)
\item \code{year >= 2013}  uses distribution z-score of 3.09 (99.9\% of the distribution)
}
}
\note{
Edited by CDMartinez 29 Dec 15
}
\examples{
OGasmt <- conventionalAssessment(auMC = 5000, auType = "Oil", auCharge = 1, auRocks = 1,
auTiming = 1, auFieldNum = c(1, 2, 10), auFieldSize = c(.5, .8, 10), auGOR = c(200, 400, 600),
auNGLGR = c(35, 85, 115), year = 2013)
}
\references{
Schmoker, James W, and T R Klett, 2005, U.S. Geological Survey Assessment Concepts
for Conventional Petroleum Accumulations. In Petroleum Systems and Geologic Assessment of Oil and Gas in the
Uinta-Piceance Province, Utah and Colorado, U.S. Geological Survey Digital Data Series DDS-69-D, 9 pages.
}
\author{
Created by CDMartinez 28 Dec 15
}
