% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsContinuousAccumulationsMatrix.R
\name{convertAcre2sqMeter}
\alias{convertAcre2sqMeter}
\title{Unit conversion}
\usage{
convertAcre2sqMeter(EA)
}
\arguments{
\item{EA}{Oil or gas assessment list}
}
\value{
Oil or gas assessment list with area variables in square meter units.
}
\description{
Convert the oil or gas assessment outputs to square meter areas instead of acres.
}
\note{
Edited by CDMartinez  15 Aug 16
}
\examples{
OGasmt <- continuousAssessment(auMC = 50000,
auType = 'Gas',
auProbability = 1,
auAreaProductive = c(2220000,2800000,3742000),
auAreaDrainage = c(10,20,40),
auPercAreaUntested = c(93,96,99),
auPercAreaSweet = c(100,100,100),
auPercFutureSS = c(80,92,94),
auEURss = c(0.15,0.4,0.65),
auLGR = c(.08,.5,1),
year = 2012)
round(continuousAssessmentSummary(OGasmt$risked))
OGmeters = convertAcre2sqMeter(OGasmt)
round(continuousAssessmentSummary(OGmeters$risked))
}
\author{
Created by CDMartinez 15 Aug 16
}
