% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsContinuousAccumulationsMatrix.R
\name{enforceRankCorrelation}
\alias{enforceRankCorrelation}
\title{Rank Correlation}
\usage{
enforceRankCorrelation(auMC, auAreaDrainage, auEURss, auEURns = NULL, year)
}
\arguments{
\item{auMC}{number of MC iterations to perform}

\item{auAreaDrainage}{Uncertainty about average drainage area of wells: \code{c(min,mode,max)} [acres]}

\item{auEURss}{Uncertainty about sweet spot average EUR (MMBO for oil; BCFG for gas): \code{c(min,med,max)} [MMBO or BCFG]}

\item{auEURns}{Uncertainty about non-sweet spot average EUR (MMBO for oil; BCFG for gas): \code{c(min,med,max)} [MMBO or BCFG]}

\item{year}{Year [XXXX] of factsheet publication of assessment numbers.}
}
\value{
Matrix of correlated variables in each row. Individual columns are \code{DrainageArea},
\code{EURsweetSpot},\code{EURnonSweetSpot}.
}
\description{
Implements rank correlation method proposed by Iman & Conover (1982).
}
\details{
Assumes correlation matrix \code{C} required by USGS Continuous Assessment
methodology between mean drainage area, sweet spot EUR, and non-sweet spot EUR
\code{C = rbind(c(1,.5,.5),c(0.5,1,0),c(0.5,0,1))}.
As of July 2016, the standard 'z-score' for the EUR distribution is implemented as 2.326 (99\%).
Assigning \code{year} <= 1 will assume zscore of 3.09 (99.9\%) as used in
\code{\link{conventionalAssessment}} EUR distributions since approximately 2013.
}
\note{
Edited by CDMartinez 08 Dec 15
}
\examples{
enforceRankCorrelation(auMC = 50000,
auAreaDrainage = c(10,20,40),
auEURss=c(0.15,0.4,0.65),year=2016)
}
\references{
Iman, Ronald L, and W J Conover, 1982, A Distribution-free approach to inducing rank
correlation among input variables. Communications in statistics - Simulation and Computation 11(3), 311-34.
}
\author{
Created by CDMartinez 10 Nov 15
}
