% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{getMaxOutline}
\alias{getMaxOutline}
\title{Get maximum productive area shapefile outline}
\usage{
getMaxOutline(EA, auOutline)
}
\arguments{
\item{EA}{Oil/gas assessment list}

\item{auOutline}{A shapefile}
}
\value{
A shapefile with same inShape projection
}
\description{
Returns maximum productive area outline using basic shape
scaling using AU shape centroid.
}
\note{
Edited by CDMartinez 14 Dec 15
}
\examples{
library(sp)
points <- rbind(c(250,250),c(250,1750),c(1750,1750),c(1750,250),c(250,250))
shape <- SpatialPolygons(list(Polygons(list(Polygon(points)), 'auOutline')))

set.seed(46)
OGasmt <- continuousAssessment(auMC = 5,
auType = 'Gas',
auProbability = 1,
auAreaProductive = c(100,400,800),
auAreaDrainage = c(10,20,40),
auPercAreaUntested = c(93,96,99),
auPercAreaSweet = c(100,100,100),
auPercFutureSS = c(20,40,50),
auEURss = c(0.15,0.4,0.65),
auLGR = c(.08,.5,1),
year = 2016)

OGasmt <- convertAcre2sqMeter(OGasmt)
maxOutline <- getMaxOutline(OGasmt, auOutline = shape)
plot(maxOutline, axes=TRUE, border = 'red')
plot(shape, add = TRUE, border = 'black')
}
\author{
Created by CDMartinez 03 Dec 15
}
