% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{importRaster}
\alias{importRaster}
\title{Imports raster file}
\usage{
importRaster(pathLoc, filename)
}
\arguments{
\item{pathLoc}{The raster path directory}

\item{filename}{The file name of the raster to import}
}
\value{
Raster
}
\description{
Wrapper function for importing a raster given the path and file name.
}
\details{
Include extension in file name. \code{myRaster <-
importRaster(pathLoc = 'path', filename = 'name.tiff')}
}
\note{
Edited by CDMartinez 17 Dec 15
}
\author{
Created by CDMartinez 17 Dec 15
}
