% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleIterationFunctions.R
\name{placePads}
\alias{placePads}
\title{Place pads for one set of sampled oil and gas parameters}
\usage{
placePads(simList, entry)
}
\arguments{
\item{simList}{List returned from \code{\link{prepareSimDistributions}}.}

\item{entry}{The index of constant values to extract from \code{simList}.}
}
\value{
Returns a list with the following items: rDrained,mPads,xyPadCenter,mcOutline,mcSweet
}
\description{
Places pads for a single Monte Carlo iteration.
}
\details{
Optimal use of this single iteration function is
utilizing a preset function \code{\link{simIteration}}
}
\note{
Edited by CDMartinez  14 Feb 17. Replaces runSingleIterationSweet
which is original implementation.
}
\examples{
library(raster)
set.seed(46)
OGasmt <- continuousAssessment(auMC = 5,
auType = 'Gas',
auProbability = 1,
auAreaProductive = c(100,400,800),
auAreaDrainage = c(10,20,40),
auPercAreaUntested = c(93,96,99),
auPercAreaSweet = c(100,100,100),
auPercFutureSS = c(20,40,50),
auEURss = c(0.15,0.4,0.65),
auLGR = c(.08,.5,1),
year = 2016)

OGasmt <- convertAcre2sqMeter(OGasmt)

rBase <- raster(resolution = c(10,10), xmn = 0, xmx = 2000, ymn = 0, ymx = 2000)
values(rBase) <- seq(1:40000)

points <- rbind(c(250,250),c(250,1750),c(1750,1750),c(1750,250),c(250,250))
shape <- SpatialPolygons(list(Polygons(list(Polygon(points)), 'auOutline')))

plot(rBase, xlim = c(0,2000), ylim = c(0,2000))
lines(shape)

spatialPrep <- prepareSimSpatial(surfaceRaster = rBase, shape, OGasmt)
distributionPrep <- prepareSimDistributions(spatialPrep,wellsPerPad = 3,
padArea = 500, EA = OGasmt, numIterations=5)

disturbance <- placePads(distributionPrep, 5)
rPads = raster(disturbance$mPads)
extent(rPads) = extent(spatialPrep$rGrid)
plot(rPads)
points(disturbance$xyPadCenter, col = 'white', pch = '.')
legend('topleft', legend = 'Gas Pad', col = terrain.colors(12), pch = 15)
}
\author{
Created by CDMartinez 15 Mar 16
}
