% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsContinuousAccumulationsMatrix.R
\name{plotLogNormalDistribution}
\alias{plotLogNormalDistribution}
\title{Plot Log-normal Distribution}
\usage{
plotLogNormalDistribution(plotName, plotVarSample, plotVarParams, bins)
}
\arguments{
\item{plotName}{Name of data being plotted (e.g. 'Drainage Area')}

\item{plotVarSample}{Sampled data vector}

\item{plotVarParams}{Log-normal distribution parameters c(min,mode,max)}

\item{bins}{Number of bins to use for distribution}
}
\value{
Plot in default viewer
}
\description{
Plots a density histogram of random sample with specified log-normal distribution overlaid.
}
\details{
Sample distribution bin = 1000 by default
}
\note{
Edited by CDMartinez 30 Nov 15
}
\examples{
OGasmt = continuousAssessment(auMC = 50000,
auType = 'Gas',
auProbability = 1,
auAreaProductive = c(100,1000,1000000),
auAreaDrainage = c(10,20,40),
auPercAreaUntested = c(93,96,99),
auPercAreaSweet = c(100,100,100),
auPercFutureSS = c(80,92,94),
auEURss = c(0.15,0.4,0.65),
auLGR = c(.08,.5,1),
year = 2016)

plotLogNormalDistribution('Average EUR',OGasmt$risked$EURsweet, c(0.15,0.4,0.65),100)
}
\author{
Created by CDMartinez 30 Nov 15
}
