% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsContinuousAccumulationsMatrix.R
\name{plotTriDistribution}
\alias{plotTriDistribution}
\title{Plot Triangle Distribution}
\usage{
plotTriDistribution(plotName, plotVarSample, plotVarParams, bins)
}
\arguments{
\item{plotName}{Name of data being plotted (e.g. 'Drainage Area')}

\item{plotVarSample}{Sampled data vector}

\item{plotVarParams}{Triangle distribution parameters c(min,mode,max)}

\item{bins}{Number of bins to use for distribution}
}
\value{
Plot in default viewer
}
\description{
Plots a density histogram of random sample with specified triangle distribution overlaid.
}
\details{
Sample distribution bin width = 1 for unrisked or (max-min)/4 for risked
}
\note{
Edited by CDMartinez 30 Nov 15
}
\examples{
OGasmt = continuousAssessment(auMC = 50000,
auType = 'Gas',
auProbability = 1,
auAreaProductive = c(100,1000,1000000),
auAreaDrainage = c(10,20,40),
auPercAreaUntested = c(93,96,99),
auPercAreaSweet = c(100,100,100),
auPercFutureSS = c(80,92,94),
auEURss = c(0.15,0.4,0.65),
auLGR = c(.08,.5,1),
year = 2016)

plotTriDistribution('Drainage Area',OGasmt$risked$DrainageArea, c(10,20,40),50)
}
\author{
Created by CDMartinez 30 Nov 15
}
