% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impactFunctions.R
\name{prepareChangeShape}
\alias{prepareChangeShape}
\title{Prepare Shapefile for Change Comparison}
\usage{
prepareChangeShape(shape, shapeAttribute = NULL, spatialList)
}
\arguments{
\item{shape}{The shapefile to prepare for simulation. Shapefile must have an attribute
representing the surface classifications.}

\item{shapeAttribute}{String providing the field name of the attribute to be used
for classification.}

\item{spatialList}{The list returned from \code{\link{prepareSimSpatial}} containing the base
raster resolution to be used for conversion.}
}
\value{
List returned from \code{\link{prepareChangeRaster}} with the matrix version of raster with required spatial extents,
number of classifications present, integer classification values
}
\description{
Prepares a shapefile values and parameters for change comparison
}
\note{
Edited by CDMartinez 8 Feb 17
}
\examples{
library(raster)
OGasmt = continuousAssessment(auMC = 50000,
auType = 'Gas',
auProbability = 1,
auAreaProductive = c(2220000,2800000,3742000),
auAreaDrainage = c(10,20,40),
auPercAreaUntested = c(93,96,99),
auPercAreaSweet = c(100,100,100),
auPercFutureSS = c(80,92,94),
auEURss = c(0.15,0.4,0.65),
auLGR = c(.08,.5,1),
year = 2012)

rBase = raster(resolution = c(10,10), xmn = 0, xmx = 100, ymn = 0, ymx = 100)
values(rBase) = seq(1:100)
points = rbind(c(30, 40), c(30, 50), c(40, 50), c(40, 60), c(50, 60), c(60, 60),
c(60, 50), c(50, 50), c(50, 40), c(40, 40), c(30, 40))
shape = spPolygons(points)

plot(rBase)
lines(shape)

prepSpatial = prepareSimSpatial(rBase,shape, OGasmt)

criticalArea <- SpatialPolygons(list(
Polygons(list(Polygon(rbind(c(0,500),c(1000,2000),c(1500,2000),c(0,0),c(0,500)))), 'A'),
Polygons(list(Polygon(rbind(c(1300,700),c(1100,950),c(900,700),c(1100,450),c(1300,700)))),'B')))
criticalAreaShape <- SpatialPolygonsDataFrame(criticalArea, data.frame(
Years = rbind(3,5), Designation = rbind('Wildlife Corridor', 'Nest Site'),
row.names = c('A','B')), match.ID = TRUE)
prepCriticalArea <- prepareChangeShape(shape = criticalAreaShape,
shapeAttribute = 'Designation', spatialList = prepSpatial)
}
\author{
Created by CDMartinez 21 Oct 16
}
