% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{prepareRusle}
\alias{prepareRusle}
\title{Prepare RUSLE inputs}
\usage{
prepareRusle(spatialList, R, K, LS, C, shapes = NULL, shapeAttribute = NULL)
}
\arguments{
\item{spatialList}{Output generated by \code{\link{prepareSimSpatial}}}

\item{R}{Rainfall erosivity factor raster}

\item{K}{Soil erodibility factor raster}

\item{LS}{Topographic slope length factor raster}

\item{C}{Vegetative cover factor raster}

\item{shapes}{Optional shapefile. If provided, simulation will return total soil loss relative to baseline as
aggregated to input polygons specified by shapefile. \code{shapes} will invoke call to \code{\link{prepareChangeShape}}}

\item{shapeAttribute}{Optional attribute. If provided along with \code{shapes} shapefile,
will invoke call to \code{\link{prepareChangeShape}}}
}
\value{
List containing inputs used for soil loss simulation, \code{\link{rusle}}. List contains
baseline input (R*K*LS*C) and C as matrices and optional input shapes for aggregated output.
}
\description{
Calculates base soil loss given initial conditions for the factors.
}
\note{
Edited by CDMartinez 2 Feb 17
}
\examples{
library(raster)
set.seed(46)
OGasmt <- continuousAssessment(auMC = 5,
auType = 'Gas',
auProbability = 1,
auAreaProductive = c(100,400,800),
auAreaDrainage = c(10,20,40),
auPercAreaUntested = c(93,96,99),
auPercAreaSweet = c(100,100,100),
auPercFutureSS = c(20,40,50),
auEURss = c(0.15,0.4,0.65),
auLGR = c(.08,.5,1),
year = 2016)

OGasmt <- convertAcre2sqMeter(OGasmt)

rBase <- raster(resolution = c(10,10), xmn = 0, xmx = 2000, ymn = 0, ymx = 2000)
values(rBase) <- sample(1:10, 40000, replace = TRUE)

points <- rbind(c(250,250),c(250,1750),c(1750,1750),c(1750,250),c(250,250))
shape <- SpatialPolygons(list(Polygons(list(Polygon(points)), 'auOutline')))

plot(rBase, xlim = c(0,2000), ylim = c(0,2000))
lines(shape)

prepSpatial <- prepareSimSpatial(surfaceRaster = rBase, shape, OGasmt)
dxdy <- 400

# Slope-length factor
tempI <- matrix(complex( real = rep(seq(0.4, .47, length.out = dxdy), each = dxdy ),
imag = rep(seq(.3, .42, length.out = dxdy), dxdy)), ncol= dxdy, nrow = dxdy)
tempZ <- 0
for(k in 1:20){tempZ <- tempZ^2+tempI}
rLS <- raster(exp(-abs(tempZ))*20)
extent(rLS) = extent(prepSpatial$rGrid)

# Vegetative cover factor
rC <- raster(matrix(sample(seq(.001, .05, length.out = 7),
size = dxdy, replace = TRUE), nrow = sqrt(dxdy)), xmn = -500,
xmx = 2500, ymn = -500, ymx = 2500)

# Soil erodibility factor
rK <- raster(matrix(sample(seq(.01, .7, length.out = 15),
size = dxdy, replace = TRUE), nrow = sqrt(dxdy)), xmn = -500,
xmx = 2500, ymn = -500, ymx = 2500)

# Rainfall erosivity
rR <- raster(matrix(rep(10, dxdy), nrow = sqrt(dxdy)),
xmn = -500, xmx = 2500, ymn = -500, ymx = 2500)

# Soil Loss will be aggregated across entire area
prepRusle <- prepareRusle(prepSpatial, R = rR, K = rK, LS = rLS, C = rC)
}
\author{
Created by CDMartinez 21 Oct 16
}
