% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{prepareSimDistributions}
\alias{prepareSimDistributions}
\title{Prepare simulation inputs related to the oil and gas variables}
\usage{
prepareSimDistributions(simSpatial, wellsPerPad, padArea, EA,
  numIterations = NULL)
}
\arguments{
\item{simSpatial}{List returned from \code{\link{prepareSimSpatial}}}

\item{wellsPerPad}{Number of wells per pad. May be either a constant integer or a vector of integers
the same length as the number of simulation iterations (\code{numIterations}) to perform.}

\item{padArea}{Area of each pad. Units should be consistent with \code{EA} units.
May be either a constant integer or a vector of integers the same length as the number of
simulation iterations (\code{numIterations}) to perform.}

\item{EA}{Oil or gas assessment list returned from \code{\link{continuousAssessment}}. Suggest
converting to square meters using \code{\link{convertAcre2sqMeter}} prior to using this function.}

\item{numIterations}{Number of MC iterations to perform}
}
\value{
List for use in spatial simulation [area units must be square meters]
}
\description{
Calculates constant variables for each iteration prior to simulation.
}
\details{
The \code{numIterations} to perform should be less than or equal to the number of
Monte Carlo iterations used to generate the energy
assessment list \code{EA}. If less than the number of energy assessments iterations available,
random sampling will determine the
energy assessment combinations to use.
}
\note{
Edited by CDMartinez  15 Nov 16
}
\examples{
library(raster)
OGasmt = continuousAssessment(auMC = 50000,
auType = 'Gas',
auProbability = 1,
auAreaProductive = c(2220000,2800000,3742000),
auAreaDrainage = c(10,20,40),
auPercAreaUntested = c(93,96,99),
auPercAreaSweet = c(100,100,100),
auPercFutureSS = c(80,92,94),
auEURss = c(0.15,0.4,0.65),
auLGR = c(.08,.5,1),
year = 2012)

rBase = raster(resolution = c(10,10), xmn = 0, xmx = 100, ymn = 0, ymx = 100)
values(rBase) = seq(1:100)
points = rbind(c(30, 40), c(30, 50), c(40, 50), c(40, 60), c(50, 60), c(60, 60),
c(60, 50), c(50, 50), c(50, 40), c(40, 40), c(30, 40))
shape = spPolygons(points)

plot(rBase)
lines(shape)

spatialPrep = prepareSimSpatial(rBase,shape, OGasmt)
distributionPrep = prepareSimDistributions(spatialPrep,wellsPerPad = 3,
padArea = 100, EA = OGasmt, numIterations=5)
}
\author{
Created by CDMartinez 15 Sept 16
}
