% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{prepareSimSpatial}
\alias{prepareSimSpatial}
\title{Prepare simulation inputs related to the spatial extents}
\usage{
prepareSimSpatial(surfaceRaster, auOutline, EA, sweetOutline = NULL,
  affineMatrix = NULL)
}
\arguments{
\item{surfaceRaster}{Surface raster with desired resolution for modeling impacts}

\item{auOutline}{The AU outline shapefile.}

\item{EA}{Oil or gas continuous assessment from \code{\link{continuousAssessment}} and converted
to square meters using \code{\link{convertAcre2sqMeter}}.}

\item{sweetOutline}{Optional. The Sweet Spot shapefile. If not provided, \code{\link{continuousAssessment}}
should likewise be an all Sweet Spot assessment.}

\item{affineMatrix}{Optional. The affine matrix to  use in \code{\link{scaleShapeTransformation}}.}
}
\value{
List for use in spatial simulation. Do not alter output from this function, as the list
is required for use of other functions such as \code{\link{prepareSimDistributions}} or \code{\link{prepareRusle}}.
}
\description{
Generate spatial variables needed for simulation. The \code{surfaceRaster} resolution
will be used for all calculated impacts and defines the smallest resolution of placing well pads.
}
\details{
If \code{affineMatrix} is not provided, default scaling of the AU extent will occur using
the shapefile centroid and \code{\link{scaleShape}}.
}
\note{
Edited by CDMartinez  6 April 17
}
\examples{
library(raster)
OGasmt = continuousAssessment(auMC = 50000,
auType = 'Gas',
auProbability = 1,
auAreaProductive = c(2220000,2800000,3742000),
auAreaDrainage = c(10,20,40),
auPercAreaUntested = c(93,96,99),
auPercAreaSweet = c(100,100,100),
auPercFutureSS = c(80,92,94),
auEURss = c(0.15,0.4,0.65),
auLGR = c(.08,.5,1),
year = 2012)

rBase = raster(resolution = c(10,10), xmn = 0, xmx = 100, ymn = 0, ymx = 100)
values(rBase) = seq(1:100)
points = rbind(c(30, 40), c(30, 50), c(40, 50), c(40, 60), c(50, 60), c(60, 60),
c(60, 50), c(50, 50), c(50, 40), c(40, 40), c(30, 40))
shape = spPolygons(points)

plot(rBase)
lines(shape)

spatialPrep <- prepareSimSpatial(rBase,shape, OGasmt)
}
\author{
Created by CDMartinez 15 Sept 16
}
