% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{scaleShape}
\alias{scaleShape}
\title{Scales shapefile.}
\usage{
scaleShape(inShape, newArea)
}
\arguments{
\item{inShape}{A shapefile}

\item{newArea}{A number for the desired area}
}
\value{
A scaled shapefile with specified area and projection of inShape
}
\description{
Scales input shapefile to desired area assuming stationary centroid.
Units must be the same.
}
\details{
For more complex shape scaling and transformations,
see \code{\link{scaleShapeTransformation}}
}
\note{
Edited by CDMartinez 14 Dec 15
}
\examples{
library(sp)
points <- rbind(c(250,250),c(250,1750),c(1750,1750),c(1750,250),c(250,250))
shape <- SpatialPolygons(list(Polygons(list(Polygon(points)), 'auOutline')))
shape2 <- scaleShape(shape, 2000000)
plot(shape, axes = TRUE)
plot(shape2, add = TRUE)
}
\author{
Created by CDMartinez 03 Dec 15
}
