% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{scaleShape2Shape}
\alias{scaleShape2Shape}
\title{Scale shapefile within shapefile}
\usage{
scaleShape2Shape(innerShape, outerShape, innerArea)
}
\arguments{
\item{innerShape}{A shapefile to be scaled}

\item{outerShape}{A shapefile limiting extent of shape scaling}

\item{innerArea}{A number for the desired area of shape output}
}
\value{
A scaled shapefile with specified area and same projection as input.
}
\description{
Scales input shape to desired area within extent of outer shape.
Units must be the same.
}
\note{
Edited by CDMartinez 15 Mar 16
}
\examples{
library(sp)
points <- rbind(c(250,250),c(250,1750),c(1750,1750),c(1750,250),c(250,250))
ShapeOuter <- SpatialPolygons(list(Polygons(list(Polygon(points)), 'auOutline')))
triPoints <- rbind(c(500,500),c(1000,1500),c(1500,500),c(500,500))
shape <- SpatialPolygons(list(Polygons(list(Polygon(triPoints)), 'maxOutline')))
shape2 <- scaleShape2Shape(innerShape = shape, outerShape = ShapeOuter, innerArea = 1500000)
plot(ShapeOuter, axes = TRUE)
plot(shape, add = TRUE, border = 'black')
plot(shape2, add = TRUE, border = 'red')
}
\author{
Created by CDMartinez 11 Mar 16
}
