% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{scaleShapeTransformation}
\alias{scaleShapeTransformation}
\title{Transforms shapefile using an affine transformation.}
\usage{
scaleShapeTransformation(inShape, newArea, percent = NULL,
  orientation = NULL, affine = NULL)
}
\arguments{
\item{inShape}{A shapefile}

\item{newArea}{A number for the desired area of the transformed shape}

\item{percent}{The decimal percent ratio to transform the shapefile}

\item{orientation}{The direction to transform the shapefile}

\item{affine}{A 2D affine matrix to apply to shapefile}
}
\value{
A scaled shapefile with specified area and projection of \code{inShape}
}
\description{
Scales input shapefile to desired area. Units must be the same.
}
\details{
Supply either the \code{affine} matrix or the \code{percent} and
\code{orientation} that will be used to transform the shapefile. If \code{orientation}
is one of the cardinal directions then \code{percent} is the decimal ratio for
shapefile elongation in the specified direction relative to the perpendicular
direction axis. If \code{orientation} is one of the ordinal directions, then
\code{percent} is the decimal ratio for shear transformation of the shapefile
in the specified direction relative to the perpendiuclar direction axis.
}
\note{
Edited by CDMartinez 14 June 16
}
\examples{
library(sp)
points <- rbind(c(250,250),c(250,1750),c(1750,1750),c(1750,250),c(250,250))
shape <- SpatialPolygons(list(Polygons(list(Polygon(points)), 'auOutline')))
shape2 <- scaleShapeTransformation(inShape = shape,newArea = 1500000,
percent = .05, orientation='SW')
shape3 <- scaleShapeTransformation(inShape = shape,newArea = 1500000,
affine = rbind(c(.2,1),c(1,.2)))
plot(shape, axes = TRUE)
plot(shape2, add = TRUE, border = 'red')
plot(shape3, add = TRUE, border = 'blue')
}
\author{
Created by CDMartinez 11 May 16
}
