% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{scaleSquareShape}
\alias{scaleSquareShape}
\title{Scales square shapefile}
\usage{
scaleSquareShape(inShape, newArea)
}
\arguments{
\item{inShape}{A square shapefile}

\item{newArea}{A number for the desired area}
}
\value{
A scaled square shapefile with specified area and projection of \code{inShape}
}
\description{
Scales input square to desired area. Units must be the same.
Suggest using \code{\link{scaleShape}} instead.
}
\note{
Edited by CDMartinez 14 Dec 15
}
\examples{
library(sp)
points <- rbind(c(250,250),c(250,1750),c(1750,1750),c(1750,250),c(250,250))
shape <- SpatialPolygons(list(Polygons(list(Polygon(points)), 'auOutline')))
shape2 <- scaleSquareShape(inShape = shape,newArea = 1500000)
plot(shape, axes = TRUE)
plot(shape2, add = TRUE, border = 'red')
}
\author{
Created by CDMartinez 03 Dec 15
}
