% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{simAggregate}
\alias{simAggregate}
\title{Aggregate output distributions}
\usage{
simAggregate(simOutput, rasterSurfaceInputs = NULL, soilLoss = NULL)
}
\arguments{
\item{simOutput}{List returned from apply call to \code{\link{simIteration}}.}

\item{rasterSurfaceInputs}{Optional. List of surface inputs to \code{\link{simIteration}}.
If provided, names will be automatically ported to the output.}

\item{soilLoss}{Optional. List returned from \code{\link{prepareRusle}}
If provided, names will be automatically ported to the output.}
}
\value{
List of aggregated impact distributions.
}
\description{
Extracts alike distribution data from list simulation output from
\code{\link{simIteration}} that was used in conjuction with a list apply function
such as \code{\link{lapply}}, \code{\link{mclapply}}, or \code{\link{parLapply}}.
}
\note{
Edited by CDMartinez 23 March 17
}
\author{
Created by CDMartinez 23 March 17
}
