% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{simIteration}
\alias{simIteration}
\title{Commands for a single iteration}
\usage{
simIteration(X, simDistList, rasterSurfaceInputs, roads = NULL,
  soilLoss = NULL, totalLength = TRUE, roadWidth = NULL,
  cellProportion = NULL)
}
\arguments{
\item{X}{Integer specifying iteration index in \code{simDistList}}

\item{simDistList}{The list returned by \code{\link{prepareSimDistributions}}}

\item{rasterSurfaceInputs}{Optional. List containing all surface rasters prepared using
\code{\link{prepareChangeShape}} or \code{\link{prepareChangeRaster}}}

\item{roads}{Optional matrix or vector. If vector is provided, road length is calculated by 
drawing a road length for each simulated pad from the provided vector of values. If matrix is provided, 
road length is calculated using distance between simulated pad and existing road network. 
The matrix format should be two column matrix with easting, northing vertices of known road network.}

\item{soilLoss}{Optional. The list returned by \code{\link{prepareRusle}}}

\item{totalLength}{Optional. Argument passed to \code{\link{makeRoads}}. \code{TRUE}
or \code{FALSE} boolean for returning total road length.
For use with \code{placeRoads}, leave as default value of \code{TRUE}.}

\item{roadWidth}{Optional. Argument passed to \code{\link{placeRoads}}.}

\item{cellProportion}{Optional. Argument passed to \code{\link{placeRoads}}.}
}
\value{
List for each iteration containing a outputs for each of the three inputs.
}
\description{
Optional utility function for a single iteration: computes surface disturbance,
road disturbance, and soil loss.
}
\note{
Edited by CDMartinez 23 Feb 17
}
\author{
Created by CDMartinez 15 February 17
}
