% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impactFunctions.R
\name{surfaceBase}
\alias{surfaceBase}
\title{Assign as base raster}
\usage{
surfaceBase(rBase)
}
\arguments{
\item{rBase}{The input raster with desired resolution and spatial extent.}
}
\value{
A list with the raster data as a matrix and spatial extent object.
}
\description{
The input raster defines the resolution for all outputs modeled
at the land surface. Conversion of the raster to the matrix format used by other
simulation functions also occurs.
}
\details{
Do not change the names of items included in the list; these will be used
for other functionality.
}
\note{
Edited by CDMartinez 28 June 16
}
\examples{
library(raster)
rLogo = raster(system.file("external/rlogo.grd", package="raster"))
plot(rLogo)
logoList = surfaceBase(rLogo)
summary(logoList)
logoList$rectBB
image(logoList$mBase)
}
\author{
Created by CDMartinez 28 June 16
}
