% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impactFunctions.R
\name{surfaceConvert}
\alias{surfaceConvert}
\title{Convert raster to matrix}
\usage{
surfaceConvert(rSurface)
}
\arguments{
\item{rSurface}{The input raster with resolution and spatial extent associated
with the base raster returned by \code{\link{surfaceBase}}.}
}
\value{
A list with the raster data as a matrix and spatial extent object.
}
\description{
Conversion of a raster to the matrix format used by other
simulation functions.
}
\note{
Edited by CDMartinez 28 June 16
}
\examples{
library(raster)
rLogo = raster(system.file("external/rlogo.grd", package="raster"))
plot(rLogo)
mLogo = surfaceConvert(rLogo)
image(mLogo)
}
\author{
Created by CDMartinez 28 June 16
}
