% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impactFunctions.R
\name{surfaceResample}
\alias{surfaceResample}
\title{Resample raster}
\usage{
surfaceResample(rSurface, rBase)
}
\arguments{
\item{rSurface}{Raster to be resampled.}

\item{rBase}{Raster with desired resolution and extent.}
}
\value{
Raster sampled from data in \code{rSurface} with resolution and extent of \code{rBase}
}
\description{
Resamples one raster to the resolution of the second raster.
}
\note{
Edited by CDMartinez 28 June 16
}
\examples{
library(raster)
rLogo = raster(system.file("external/rlogo.grd", package="raster"))
rBase = raster(resolution = c(0.5,0.5), xmn = 0, xmx = 101, ymn = 0, ymx = 77)
rLogoFine = surfaceResample(rLogo, rBase)
plot(rLogoFine)
plot(rLogo)
}
\author{
Created by CDMartinez 28 June 16
}
